Add webmaster's choice font selection.
Clean up the radio selection functions to simplify adding additional
options and maintaining existing ones.

The newly added Charter font, should hopefully be lazily-loaded only
if a user selects the webmaster's choice option. It looks to be about
72KB in size, so much bigger than rest of the index page, but not
outlandish. Tweak the CSS to group the fonts at the end now.