Add missing images and animation warning.
These images weren't indexed by the Internet Archive since they are
only referenced within JavaScript that is not necessarily called. They
are related to the homepage puzzle.

My added skip button gets some more help text to indicate one has to
wait for the animation to play to access the menu. I disable it after
the first click, since it appears to mess with the puzzle/menu state.