Add alternate Dobos Torte recipe, tighten language.
Add a seemingly solid YouTube video I found, and tweak some language
seeking improved clarity. According to the video, the butter was for
the knife during the caramel cutting stage. I wasn't sure what it was
for anyway from the initial recipe, so I've moved it to refer to that.