Organize colorscheme with CSS variables.
Use CSS variables to simplify colorscheme management and reduce
duplicated effort, like with max-width. Thanks to Alan Hogan for his
instructive CSS template that I adapted and linked to in my stylesheet.

Didn't change the colors that much with this commit, but it lays a
good foundation for colorscheme customization in the future.